pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "lazypaper"
version = "0.0.0"
description = "Daily journal-article email digest"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "beautifulsoup4>=4.12.0",
    "dnspython>=2.6.0",
    "feedparser>=6.0.11",
    "requests>=2.31.0",
    "resend>=2.0.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
markers = [
    "network: fetches live RSS and Europe PMC (requires internet)",
]
